refactored the code to have separate folder for enums
