+++ backend/app/data/dummy_data.py
@@ -8,34 +8,34 @@ from supabase import AuthError
 
 from app.config import settings
 from app.database import get_supabase_client
+from app.enums.account_role import AccountRole
+from app.enums.confidence_area import ConfidenceArea
+from app.enums.config_type import ConfigType
+from app.enums.conversation_scenario_status import ConversationScenarioStatus
+from app.enums.difficulty_level import DifficultyLevel
+from app.enums.experience import Experience
+from app.enums.feedback_status import FeedbackStatus
+from app.enums.goal import Goal
+from app.enums.language import LanguageCode
+from app.enums.preferred_learning_style import PreferredLearningStyle
+from app.enums.professional_role import ProfessionalRole
+from app.enums.scenario_preparation_status import ScenarioPreparationStatus
+from app.enums.session_status import SessionStatus
+from app.enums.speaker import SpeakerType
 from app.interfaces import MockUserIdsEnum
 from app.models import LiveFeedback
 from app.models.admin_dashboard_stats import AdminDashboardStats
-from app.models.app_config import AppConfig, ConfigType
+from app.models.app_config import AppConfig
 from app.models.conversation_category import ConversationCategory
-from app.models.conversation_scenario import (
-    ConversationScenario,
-    ConversationScenarioStatus,
-    DifficultyLevel,
-)
-from app.models.language import LanguageCode
+from app.models.conversation_scenario import ConversationScenario
 from app.models.review import Review
-from app.models.scenario_preparation import ScenarioPreparation, ScenarioPreparationStatus
-from app.models.session import Session, SessionStatus
-from app.models.session_feedback import (
-    FeedbackStatusEnum,
-    SessionFeedback,
-)
-from app.models.session_turn import SessionTurn, SpeakerEnum
-from app.models.user_confidence_score import ConfidenceArea, UserConfidenceScore
-from app.models.user_goal import Goal, UserGoal
-from app.models.user_profile import (
-    AccountRole,
-    Experience,
-    PreferredLearningStyle,
-    ProfessionalRole,
-    UserProfile,
-)
+from app.models.scenario_preparation import ScenarioPreparation
+from app.models.session import Session
+from app.models.session_feedback import SessionFeedback
+from app.models.session_turn import SessionTurn
+from app.models.user_confidence_score import UserConfidenceScore
+from app.models.user_goal import UserGoal
+from app.models.user_profile import UserProfile
 
 
 def get_dummy_user_profiles() -> list[UserProfile]:
@@ -442,7 +442,7 @@ def get_dummy_session_turns(
         SessionTurn(
             id=uuid4(),
             session_id=sessions[0].id,
-            speaker=SpeakerEnum.user,
+            speaker=SpeakerType.user,
             start_offset_ms=0,
             end_offset_ms=5000,
             text='Hello, how can I help you?',
@@ -453,7 +453,7 @@ def get_dummy_session_turns(
         SessionTurn(
             id=uuid4(),
             session_id=sessions[0].id,
-            speaker=SpeakerEnum.assistant,
+            speaker=SpeakerType.assistant,
             start_offset_ms=5000,
             end_offset_ms=10000,
             text="Hi! I'd like to check your schedule today. Are you available at 2 PM?",
@@ -464,7 +464,7 @@ def get_dummy_session_turns(
         SessionTurn(
             id=uuid4(),
             session_id=sessions[0].id,
-            speaker=SpeakerEnum.user,
+            speaker=SpeakerType.user,
             start_offset_ms=10000,
             end_offset_ms=15000,
             text='Sure, 2 PM works fine.',
@@ -476,7 +476,7 @@ def get_dummy_session_turns(
         SessionTurn(
             id=uuid4(),
             session_id=sessions[1].id,
-            speaker=SpeakerEnum.user,
+            speaker=SpeakerType.user,
             start_offset_ms=0,
             end_offset_ms=4000,
             text='I need assistance with my account.',
@@ -487,7 +487,7 @@ def get_dummy_session_turns(
         SessionTurn(
             id=uuid4(),
             session_id=sessions[1].id,
-            speaker=SpeakerEnum.assistant,
+            speaker=SpeakerType.assistant,
             start_offset_ms=4000,
             end_offset_ms=9000,
             text='Of course. Could you please tell me what issue you are facing?',
@@ -498,7 +498,7 @@ def get_dummy_session_turns(
         SessionTurn(
             id=uuid4(),
             session_id=sessions[1].id,
-            speaker=SpeakerEnum.user,
+            speaker=SpeakerType.user,
             start_offset_ms=9000,
             end_offset_ms=13000,
             text="I'm not able to log in since yesterday.",
@@ -509,7 +509,7 @@ def get_dummy_session_turns(
         SessionTurn(
             id=uuid4(),
             session_id=sessions[1].id,
-            speaker=SpeakerEnum.assistant,
+            speaker=SpeakerType.assistant,
             start_offset_ms=13000,
             end_offset_ms=17000,
             text='Thanks for the info. I will reset your credentials and email you shortly.',
@@ -521,7 +521,7 @@ def get_dummy_session_turns(
         SessionTurn(
             id=uuid4(),
             session_id=sessions[2].id,
-            speaker=SpeakerEnum.user,
+            speaker=SpeakerType.user,
             start_offset_ms=0,
             end_offset_ms=3000,
             text='Can you prepare the sales report by end of day?',
@@ -532,7 +532,7 @@ def get_dummy_session_turns(
         SessionTurn(
             id=uuid4(),
             session_id=sessions[2].id,
-            speaker=SpeakerEnum.assistant,
+            speaker=SpeakerType.assistant,
             start_offset_ms=3000,
             end_offset_ms=7000,
             text="Absolutely. I'll make sure to include the latest figures.",
@@ -543,7 +543,7 @@ def get_dummy_session_turns(
         SessionTurn(
             id=uuid4(),
             session_id=sessions[2].id,
-            speaker=SpeakerEnum.user,
+            speaker=SpeakerType.user,
             start_offset_ms=7000,
             end_offset_ms=10000,
             text='Good. Please double-check the Q2 numbers.',
@@ -554,7 +554,7 @@ def get_dummy_session_turns(
         SessionTurn(
             id=uuid4(),
             session_id=sessions[2].id,
-            speaker=SpeakerEnum.assistant,
+            speaker=SpeakerType.assistant,
             start_offset_ms=10000,
             end_offset_ms=13500,
             text="Will do. I'll send you a draft in a couple of hours.",
@@ -566,7 +566,7 @@ def get_dummy_session_turns(
         SessionTurn(
             id=uuid4(),
             session_id=sessions[3].id,
-            speaker=SpeakerEnum.user,
+            speaker=SpeakerType.user,
             start_offset_ms=0,
             end_offset_ms=2800,
             text="How's the client presentation coming along?",
@@ -577,7 +577,7 @@ def get_dummy_session_turns(
         SessionTurn(
             id=uuid4(),
             session_id=sessions[3].id,
-            speaker=SpeakerEnum.assistant,
+            speaker=SpeakerType.assistant,
             start_offset_ms=2800,
             end_offset_ms=6500,
             text="I've finished most of the slides. Just adding the final data now.",
@@ -588,7 +588,7 @@ def get_dummy_session_turns(
         SessionTurn(
             id=uuid4(),
             session_id=sessions[3].id,
-            speaker=SpeakerEnum.user,
+            speaker=SpeakerType.user,
             start_offset_ms=6500,
             end_offset_ms=9500,
             text='Great. Make sure the figures are up-to-date.',
@@ -599,7 +599,7 @@ def get_dummy_session_turns(
         SessionTurn(
             id=uuid4(),
             session_id=sessions[3].id,
-            speaker=SpeakerEnum.assistant,
+            speaker=SpeakerType.assistant,
             start_offset_ms=9500,
             end_offset_ms=12800,
             text="Understood. I'll send it for your review by 3 PM.",
@@ -611,7 +611,7 @@ def get_dummy_session_turns(
         SessionTurn(
             id=uuid4(),
             session_id=sessions[4].id,
-            speaker=SpeakerEnum.user,
+            speaker=SpeakerType.user,
             start_offset_ms=0,
             end_offset_ms=3200,
             text='Did you submit your leave request for next week?',
@@ -622,7 +622,7 @@ def get_dummy_session_turns(
         SessionTurn(
             id=uuid4(),
             session_id=sessions[4].id,
-            speaker=SpeakerEnum.assistant,
+            speaker=SpeakerType.assistant,
             start_offset_ms=3200,
             end_offset_ms=6000,
             text='Yes, I submitted it yesterday. Waiting for approval.',
@@ -633,7 +633,7 @@ def get_dummy_session_turns(
         SessionTurn(
             id=uuid4(),
             session_id=sessions[4].id,
-            speaker=SpeakerEnum.user,
+            speaker=SpeakerType.user,
             start_offset_ms=6000,
             end_offset_ms=9000,
             text='Alright. Make sure all your tasks are handed over before you go.',
@@ -644,7 +644,7 @@ def get_dummy_session_turns(
         SessionTurn(
             id=uuid4(),
             session_id=sessions[4].id,
-            speaker=SpeakerEnum.assistant,
+            speaker=SpeakerType.assistant,
             start_offset_ms=9000,
             end_offset_ms=12500,
             text="Of course. I'll update the team and share the status report tomorrow.",
@@ -656,7 +656,7 @@ def get_dummy_session_turns(
         SessionTurn(
             id=uuid4(),
             session_id=sessions[5].id,
-            speaker=SpeakerEnum.user,
+            speaker=SpeakerType.user,
             start_offset_ms=0,
             end_offset_ms=3500,
             text='I noticed the budget forecast is still pending. Any update?',
@@ -667,7 +667,7 @@ def get_dummy_session_turns(
         SessionTurn(
             id=uuid4(),
             session_id=sessions[5].id,
-            speaker=SpeakerEnum.assistant,
+            speaker=SpeakerType.assistant,
             start_offset_ms=3500,
             end_offset_ms=7000,
             text="Apologies for the delay. I'll finalize it by this afternoon.",
@@ -678,7 +678,7 @@ def get_dummy_session_turns(
         SessionTurn(
             id=uuid4(),
             session_id=sessions[5].id,
-            speaker=SpeakerEnum.user,
+            speaker=SpeakerType.user,
             start_offset_ms=7000,
             end_offset_ms=10500,
             text="Alright, please prioritize it. We're presenting it tomorrow.",
@@ -689,7 +689,7 @@ def get_dummy_session_turns(
         SessionTurn(
             id=uuid4(),
             session_id=sessions[5].id,
-            speaker=SpeakerEnum.assistant,
+            speaker=SpeakerType.assistant,
             start_offset_ms=10500,
             end_offset_ms=13800,
             text="Understood. I'll share it with you before the end of the day.",
@@ -701,7 +701,7 @@ def get_dummy_session_turns(
         SessionTurn(
             id=uuid4(),
             session_id=sessions[6].id,
-            speaker=SpeakerEnum.user,
+            speaker=SpeakerType.user,
             start_offset_ms=0,
             end_offset_ms=3500,
             text='I noticed the budget forecast is still pending. Any update?',
@@ -712,7 +712,7 @@ def get_dummy_session_turns(
         SessionTurn(
             id=uuid4(),
             session_id=sessions[6].id,
-            speaker=SpeakerEnum.assistant,
+            speaker=SpeakerType.assistant,
             start_offset_ms=3500,
             end_offset_ms=7000,
             text="Apologies for the delay. I'll finalize it by this afternoon.",
@@ -723,7 +723,7 @@ def get_dummy_session_turns(
         SessionTurn(
             id=uuid4(),
             session_id=sessions[6].id,
-            speaker=SpeakerEnum.user,
+            speaker=SpeakerType.user,
             start_offset_ms=7000,
             end_offset_ms=10500,
             text="Alright, please prioritize it. We're presenting it tomorrow.",
@@ -734,7 +734,7 @@ def get_dummy_session_turns(
         SessionTurn(
             id=uuid4(),
             session_id=sessions[6].id,
-            speaker=SpeakerEnum.assistant,
+            speaker=SpeakerType.assistant,
             start_offset_ms=10500,
             end_offset_ms=13800,
             text="Understood. I'll share it with you before the end of the day.",
@@ -746,7 +746,7 @@ def get_dummy_session_turns(
         SessionTurn(
             id=uuid4(),
             session_id=sessions[7].id,
-            speaker=SpeakerEnum.user,
+            speaker=SpeakerType.user,
             start_offset_ms=0,
             end_offset_ms=3500,
             text='I noticed the budget forecast is still pending. Any update?',
@@ -757,7 +757,7 @@ def get_dummy_session_turns(
         SessionTurn(
             id=uuid4(),
             session_id=sessions[7].id,
-            speaker=SpeakerEnum.assistant,
+            speaker=SpeakerType.assistant,
             start_offset_ms=3500,
             end_offset_ms=7000,
             text="Apologies for the delay. I'll finalize it by this afternoon.",
@@ -768,7 +768,7 @@ def get_dummy_session_turns(
         SessionTurn(
             id=uuid4(),
             session_id=sessions[7].id,
-            speaker=SpeakerEnum.user,
+            speaker=SpeakerType.user,
             start_offset_ms=7000,
             end_offset_ms=10500,
             text="Alright, please prioritize it. We're presenting it tomorrow.",
@@ -779,7 +779,7 @@ def get_dummy_session_turns(
         SessionTurn(
             id=uuid4(),
             session_id=sessions[7].id,
-            speaker=SpeakerEnum.assistant,
+            speaker=SpeakerType.assistant,
             start_offset_ms=10500,
             end_offset_ms=13800,
             text="Understood. I'll share it with you before the end of the day.",
@@ -953,7 +953,7 @@ def get_dummy_session_feedback(
                     ),
                 },
             ],
-            status=FeedbackStatusEnum.completed,  # Use the enum for status
+            status=FeedbackStatus.completed,  # Use the enum for status
             created_at=datetime.now(UTC),
             updated_at=datetime.now(UTC),
         ),
@@ -1031,7 +1031,7 @@ def get_dummy_session_feedback(
                     ),
                 },
             ],
-            status=FeedbackStatusEnum.completed,
+            status=FeedbackStatus.completed,
             created_at=datetime.now(UTC),
             updated_at=datetime.now(UTC),
         ),
@@ -1109,7 +1109,7 @@ def get_dummy_session_feedback(
                     ),
                 },
             ],
-            status=FeedbackStatusEnum.completed,
+            status=FeedbackStatus.completed,
             created_at=datetime.now(UTC),
             updated_at=datetime.now(UTC),
         ),
@@ -1187,7 +1187,7 @@ def get_dummy_session_feedback(
                     ),
                 },
             ],
-            status=FeedbackStatusEnum.completed,
+            status=FeedbackStatus.completed,
             created_at=datetime.now(UTC),
             updated_at=datetime.now(UTC),
         ),
@@ -1265,7 +1265,7 @@ def get_dummy_session_feedback(
                     ),
                 },
             ],
-            status=FeedbackStatusEnum.completed,  # Use the enum for status
+            status=FeedbackStatus.completed,  # Use the enum for status
             created_at=datetime.now(UTC),
             updated_at=datetime.now(UTC),
         ),
@@ -1344,7 +1344,7 @@ def get_dummy_session_feedback(
                     ),
                 },
             ],
-            status=FeedbackStatusEnum.completed,  # Use the enum for status
+            status=FeedbackStatus.completed,  # Use the enum for status
             created_at=datetime.now(UTC),
             updated_at=datetime.now(UTC),
         ),
@@ -1419,7 +1419,7 @@ def get_dummy_session_feedback(
                     ),
                 },
             ],
-            status=FeedbackStatusEnum.completed,  # Use the enum for status
+            status=FeedbackStatus.completed,  # Use the enum for status
             created_at=datetime.now(UTC),
             updated_at=datetime.now(UTC),
         ),
@@ -1492,7 +1492,7 @@ def get_dummy_session_feedback(
                     ),
                 },
             ],
-            status=FeedbackStatusEnum.completed,  # Use the enum for status
+            status=FeedbackStatus.completed,  # Use the enum for status
             created_at=datetime.now(UTC),
             updated_at=datetime.now(UTC),
         ),
