+++ backend/app/data/populate_static_data.py
@@ -2,13 +2,13 @@ import json
 import os
 from datetime import UTC, datetime
 
+from backend.app.enums.language import LanguageCode
 from sqlmodel import Session as DBSession
 from sqlmodel import col, select
 
 from app.data import create_mock_users, delete_mock_users, get_dummy_user_profiles
 from app.database import engine
 from app.models import ConversationCategory, UserProfile
-from app.models.language import LanguageCode
 
 base_dir = os.path.dirname(__file__)
 with open(os.path.join(base_dir, 'initial_prompts.json'), encoding='utf-8') as f:
