+++ backend/app/enums/__init__.py
@@ -1,0 +1,31 @@
+from app.enums.account_role import AccountRole
+from app.enums.confidence_area import ConfidenceArea
+from app.enums.config_type import ConfigType
+from app.enums.conversation_scenario_status import ConversationScenarioStatus
+from app.enums.difficulty_level import DifficultyLevel
+from app.enums.experience import Experience
+from app.enums.feedback_status import FeedbackStatus
+from app.enums.goal import Goal
+from app.enums.language import LanguageCode
+from app.enums.preferred_learning_style import PreferredLearningStyle
+from app.enums.professional_role import ProfessionalRole
+from app.enums.scenario_preparation_status import ScenarioPreparationStatus
+from app.enums.session_status import SessionStatus
+from app.enums.speaker import SpeakerType
+
+__all__ = [
+    'ConfigType',
+    'DifficultyLevel',
+    'ConversationScenarioStatus',
+    'LanguageCode',
+    'ScenarioPreparationStatus',
+    'FeedbackStatus',
+    'SpeakerType',
+    'SessionStatus',
+    'ConfidenceArea',
+    'Goal',
+    'AccountRole',
+    'ProfessionalRole',
+    'Experience',
+    'PreferredLearningStyle',
+]
