+++ backend/app/enums/account_role.py
@@ -1,0 +1,6 @@
+from enum import Enum as PyEnum
+
+
+class AccountRole(str, PyEnum):
+    user = 'user'
+    admin = 'admin'
