+++ backend/app/enums/confidence_area.py
@@ -1,0 +1,8 @@
+from enum import Enum as PyEnum
+
+
+class ConfidenceArea(str, PyEnum):
+    # Define the confidence areas as needed
+    giving_difficult_feedback = 'giving_difficult_feedback'
+    managing_team_conflicts = 'managing_team_conflicts'
+    leading_challenging_conversations = 'leading_challenging_conversations'
