+++ backend/app/enums/config_type.py
@@ -1,0 +1,8 @@
+from enum import Enum as PyEnum
+
+
+# Enum for the type column
+class ConfigType(PyEnum):
+    int = 'int'
+    string = 'string'
+    boolean = 'boolean'
