+++ backend/app/enums/conversation_scenario_status.py
@@ -1,0 +1,8 @@
+from enum import Enum as PyEnum
+
+
+# Enum for status
+class ConversationScenarioStatus(str, PyEnum):
+    draft = 'draft'
+    ready = 'ready'
+    archived = 'archived'
