+++ backend/app/enums/difficulty_level.py
@@ -1,0 +1,7 @@
+from enum import Enum as PyEnum
+
+
+class DifficultyLevel(str, PyEnum):
+    easy = 'easy'
+    medium = 'medium'
+    hard = 'hard'
