+++ backend/app/enums/experience.py
@@ -1,0 +1,9 @@
+from enum import Enum as PyEnum
+
+
+class Experience(str, PyEnum):
+    beginner = 'beginner'
+    intermediate = 'intermediate'
+    skilled = 'skilled'
+    advanced = 'advanced'
+    expert = 'expert'
