+++ backend/app/enums/feedback_status.py
@@ -1,0 +1,7 @@
+from enum import Enum as PyEnum
+
+
+class FeedbackStatus(str, PyEnum):
+    pending = 'pending'
+    completed = 'completed'
+    failed = 'failed'
