+++ backend/app/enums/language.py
@@ -1,0 +1,6 @@
+from enum import Enum as PyEnum
+
+
+class LanguageCode(str, PyEnum):
+    en = 'en'
+    de = 'de'
