+++ backend/app/enums/preferred_learning_style.py
@@ -1,0 +1,7 @@
+from enum import Enum as PyEnum
+
+
+class PreferredLearningStyle(str, PyEnum):
+    visual = 'visual'
+    auditory = 'auditory'
+    kinesthetic = 'kinesthetic'
