+++ backend/app/enums/professional_role.py
@@ -1,0 +1,8 @@
+from enum import Enum as PyEnum
+
+
+class ProfessionalRole(str, PyEnum):
+    hr_professional = 'hr_professional'
+    team_leader = 'team_leader'
+    executive = 'executive'
+    other = 'other'
