+++ backend/app/enums/scenario_preparation_status.py
@@ -1,0 +1,7 @@
+from enum import Enum as PyEnum
+
+
+class ScenarioPreparationStatus(str, PyEnum):
+    pending = 'pending'
+    completed = 'completed'
+    failed = 'failed'
