+++ backend/app/enums/session_status.py
@@ -1,0 +1,7 @@
+from enum import Enum as PyEnum
+
+
+class SessionStatus(str, PyEnum):
+    started = 'started'
+    completed = 'completed'
+    failed = 'failed'
