+++ backend/app/enums/speaker.py
@@ -1,0 +1,6 @@
+from enum import Enum as PyEnum
+
+
+class SpeakerType(str, PyEnum):
+    user = 'user'
+    assistant = 'assistant'
