+++ backend/app/models/__init__.py
@@ -1,18 +1,16 @@
 from app.models.admin_dashboard_stats import AdminDashboardStats
-from app.models.app_config import AppConfig, ConfigType
+from app.models.app_config import AppConfig
 from app.models.conversation_category import ConversationCategory
 from app.models.conversation_scenario import (
     ConversationScenario,
-    ConversationScenarioStatus,
 )
 from app.models.live_feedback_model import LiveFeedback
 from app.models.review import Review
 from app.models.scenario_preparation import (
     ScenarioPreparation,
-    ScenarioPreparationStatus,
 )
-from app.models.session import Session, SessionStatus
-from app.models.session_feedback import FeedbackStatusEnum, SessionFeedback
+from app.models.session import Session
+from app.models.session_feedback import SessionFeedback
 from app.models.session_turn import SessionTurn
 from app.models.user_confidence_score import (
     UserConfidenceScore,
@@ -23,19 +21,14 @@ from app.models.user_profile import UserProfile
 __all__ = [
     'ConversationCategory',
     'ConversationScenario',
-    'ConversationScenarioStatus',
     'Session',
-    'SessionStatus',
     'ScenarioPreparation',
-    'ScenarioPreparationStatus',
     'SessionTurn',
     'SessionFeedback',
     'UserProfile',
     'UserGoal',
-    'FeedbackStatusEnum',
     'UserConfidenceScore',
     'AppConfig',
-    'ConfigType',
     'AdminDashboardStats',
     'Review',
     'LiveFeedback',
