--- backend/app/models/language.py
@@ -1,6 +1,0 @@
-from enum import Enum
-
-
-class LanguageCode(str, Enum):
-    en = 'en'
-    de = 'de'
