+++ backend/app/models/session_feedback.py
@@ -1,5 +1,4 @@
 from datetime import UTC, datetime
-from enum import Enum
 from typing import TYPE_CHECKING, Optional
 from uuid import UUID, uuid4
 
@@ -13,11 +12,7 @@ from app.models.camel_case import CamelModel
 if TYPE_CHECKING:
     from app.models.session import Session
 
-
-class FeedbackStatusEnum(str, Enum):
-    pending = 'pending'
-    completed = 'completed'
-    failed = 'failed'
+from app.enums.feedback_status import FeedbackStatus
 
 
 class SessionFeedback(CamelModel, table=True):  # `table=True` makes it a database table
@@ -36,7 +31,7 @@ class SessionFeedback(CamelModel, table=True):  # `table=True` makes it a databa
     example_positive: list[dict] = Field(default_factory=list, sa_column=Column(JSON))
     example_negative: list[dict] = Field(default_factory=list, sa_column=Column(JSON))
     recommendations: list[dict] = Field(default_factory=list, sa_column=Column(JSON))
-    status: FeedbackStatusEnum = Field(default=FeedbackStatusEnum.pending)
+    status: FeedbackStatus = Field(default=FeedbackStatus.pending)
     created_at: datetime = Field(default_factory=lambda: datetime.now(UTC))
     updated_at: datetime = Field(default_factory=lambda: datetime.now(UTC))
