+++ backend/app/models/user_goal.py
@@ -1,5 +1,4 @@
 from datetime import UTC, datetime
-from enum import Enum
 from typing import TYPE_CHECKING, Optional
 from uuid import UUID
 
@@ -8,23 +7,11 @@ from sqlalchemy.engine import Connection
 from sqlalchemy.orm import Mapper
 from sqlmodel import Field, Relationship
 
+from app.enums.goal import Goal
 from app.models.camel_case import CamelModel
 
 if TYPE_CHECKING:
     from app.models.user_profile import UserProfile
-
-
-class Goal(str, Enum):
-    giving_constructive_feedback = 'giving_constructive_feedback'
-    managing_team_conflicts = 'managing_team_conflicts'
-    performance_reviews = 'performance_reviews'
-    motivating_team_members = 'motivating_team_members'
-    leading_difficult_conversations = 'leading_difficult_conversations'
-    communicating_organizational_change = 'communicating_organizational_change'
-    develop_emotional_intelligence = 'develop_emotional_intelligence'
-    building_inclusive_teams = 'building_inclusive_teams'
-    negotiation_skills = 'negotiation_skills'
-    coaching_mentoring = 'coaching_mentoring'
 
 
 class UserGoal(CamelModel, table=True):
