+++ backend/app/routers/session_turn_route.py
@@ -6,7 +6,7 @@ from sqlmodel import Session as DBSession
 
 from app.database import get_db_session
 from app.dependencies import require_user
-from app.models.session_turn import SpeakerEnum
+from app.enums.speaker import SpeakerType
 from app.schemas.session_turn import SessionTurnCreate, SessionTurnRead
 from app.services.session_turn_service import SessionTurnService
 
@@ -29,7 +29,7 @@ async def create_session_turn(
     service: Annotated[SessionTurnService, Depends(get_session_turn_service)],
     background_tasks: BackgroundTasks,
     session_id: UUID = Form(...),  # noqa: B008
-    speaker: SpeakerEnum = Form(...),  # noqa: B008
+    speaker: SpeakerType = Form(...),  # noqa: B008
     start_offset_ms: int = Form(...),  # noqa: B008
     end_offset_ms: int = Form(...),  # noqa: B008
     text: str = Form(...),  # noqa: B008
