+++ backend/app/schemas/scenario_prep_config.py
@@ -1,7 +1,7 @@
 from pydantic import RootModel
 
+from app.enums.language import LanguageCode
 from app.models.camel_case import CamelModel
-from app.models.language import LanguageCode
 from app.schemas.scenario_preparation import KeyConceptResponse, StringListResponse
