+++ backend/app/schemas/session_turn.py
@@ -2,14 +2,14 @@ from datetime import datetime
 from typing import Optional
 from uuid import UUID
 
+from app.enums.speaker import SpeakerType
 from app.models.camel_case import CamelModel
-from app.models.session_turn import SpeakerEnum
 
 
 # Schema for creating a new SessionTurn
 class SessionTurnCreate(CamelModel):
     session_id: UUID
-    speaker: SpeakerEnum
+    speaker: SpeakerType
     start_offset_ms: int
     end_offset_ms: int
     text: str
@@ -18,7 +18,7 @@ class SessionTurnCreate(CamelModel):
 # Schema for reading SessionTurn data
 class SessionTurnRead(CamelModel):
     id: UUID
-    speaker: SpeakerEnum
+    speaker: SpeakerType
     full_audio_start_offset_ms: int
     text: str
     ai_emotion: Optional[str] = None
