+++ backend/app/schemas/user_profile.py
@@ -1,15 +1,13 @@
 from datetime import datetime
 from uuid import UUID
 
+from app.enums.account_role import AccountRole
+from app.enums.experience import Experience
+from app.enums.language import LanguageCode
+from app.enums.preferred_learning_style import PreferredLearningStyle
+from app.enums.professional_role import ProfessionalRole
 from app.models.camel_case import CamelModel
-from app.models.language import LanguageCode
 from app.models.user_goal import Goal
-from app.models.user_profile import (
-    AccountRole,
-    Experience,
-    PreferredLearningStyle,
-    ProfessionalRole,
-)
 from app.schemas.user_confidence_score import ConfidenceScoreRead
