+++ backend/app/services/conversation_scenario_service.py
@@ -5,12 +5,13 @@ from sqlmodel import Session as DBSession
 from sqlmodel import func, select
 
 from app.database import get_db_session
+from app.enums.account_role import AccountRole
 from app.models.conversation_category import ConversationCategory
 from app.models.conversation_scenario import ConversationScenario
 from app.models.scenario_preparation import ScenarioPreparation, ScenarioPreparationStatus
 from app.models.session import Session
-from app.models.session_feedback import FeedbackStatusEnum, SessionFeedback
-from app.models.user_profile import AccountRole, UserProfile
+from app.models.session_feedback import FeedbackStatus, SessionFeedback
+from app.models.user_profile import UserProfile
 from app.schemas.conversation_scenario import (
     ConversationScenarioCreate,
     ConversationScenarioCreateResponse,
@@ -149,7 +150,7 @@ class ConversationScenarioService:
             # session  → feedback (may be zero)
             .outerjoin(SessionFeedback, SessionFeedback.session_id == Session.id)
             .where(
-                SessionFeedback.status == FeedbackStatusEnum.completed,
+                SessionFeedback.status == FeedbackStatus.completed,
             )
             .group_by(
                 ConversationScenario.id,
@@ -286,7 +287,7 @@ class ConversationScenarioService:
                 [
                     session.feedback.overall_score
                     for session in scenario.sessions
-                    if session.feedback and session.feedback.status == FeedbackStatusEnum.completed
+                    if session.feedback and session.feedback.status == FeedbackStatus.completed
                 ]
             )
             / total_sessions
