+++ backend/app/services/scenario_preparation/scenario_preparation_service.py
@@ -11,8 +11,9 @@ from sqlmodel import Session as DBSession
 from tenacity import retry, stop_after_attempt, wait_fixed
 
 from app.connections.vertexai_client import call_structured_llm
-from app.models.language import LanguageCode
-from app.models.scenario_preparation import ScenarioPreparation, ScenarioPreparationStatus
+from app.enums.language import LanguageCode
+from app.enums.scenario_preparation_status import ScenarioPreparationStatus
+from app.models.scenario_preparation import ScenarioPreparation
 from app.schemas.scenario_prep_config import ScenarioPrepConfig
 from app.schemas.scenario_preparation import (
     ChecklistRequest,
