+++ backend/app/services/session_feedback/session_feedback_llm.py
@@ -5,7 +5,7 @@ from functools import lru_cache
 from tenacity import retry, stop_after_attempt, wait_fixed
 
 from app.connections.vertexai_client import call_structured_llm
-from app.models.language import LanguageCode
+from app.enums.language import LanguageCode
 from app.schemas.session_feedback import (
     FeedbackRequest,
     GoalsAchievedCollection,
