+++ backend/app/tests/services/test_generate_training_examples.py
@@ -3,7 +3,7 @@ import os
 import string
 import unittest
 
-from app.models.language import LanguageCode
+from app.enums.language import LanguageCode
 from app.schemas.session_feedback import FeedbackRequest, SessionExamplesCollection
 from app.services.session_feedback.session_feedback_llm import generate_training_examples
