+++ backend/app/tests/services/test_scoring_service.py
@@ -5,10 +5,10 @@ from pathlib import Path
 from unittest.mock import MagicMock, patch
 from uuid import uuid4
 
+from backend.app.enums.language import LanguageCode
 from pydantic import ValidationError
 
 from app.models.conversation_scenario import ConversationScenarioStatus, DifficultyLevel
-from app.models.language import LanguageCode
 from app.models.session_turn import SpeakerEnum
 from app.schemas.conversation_scenario import (
     ConversationScenario,
